requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.6
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.0.0
Flask==1.1.2
Flask-OAuth2-Login==0.0.9
Flask-SSLify==0.1.5
idna==2.10
inventory==2.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.1.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
waitress==1.4.4
Werkzeug==1.0.1